requirements.txt
streamlit==1.28.1
google-generativeai
python-dotenv
textwrap
IPython